Fix the focus problem.
There was a problem with the way we focus the next tile, that resulted
in cases like the following that you can't focus right once you are on
window 1:

```
|          | window 2 |
| window 1 |          |
|          | window 3 |
```

I fixed this by adding a focus stack to each tile grid that contains the
previous focus actions. When a new focus action occurs that "counters"
the previous one, meaning for example:

Window 2 is focused
You focus left
You focus right

In this example the focus right action cancels the previous one.