feat(config): support border setting
The default value before this commit for the native windows border was
true with no way of disabling this.

After some usage of wwm I noticed that the border isn't really necessary
for most programms only in cases like a terminal emulator where you
usually can tell where the terminal starts via the prompt.

Because of this the default value is now false and it can be enabled
again by setting the `use_border` property at the root level to `true`.