fix: ensure that unregistering the keybindings is sync
We used to have a small delay between flagging for unregistering and
registering the keybindings, which works most of the times but when the
process is running really slow it might cause problems.

The problem was that each hot reload would close the current event loop
and open a new one. We now are reusing the same event loop during the
whole lifetime of the program, which results in synchronous unregister
and registering.